Redid routing of components and screens to implement react navigation.
