+++ package.json
@@ -11,15 +11,18 @@
   "dependencies": {
     "@react-native-firebase/app": "^20.0.0",
     "@react-native-google-signin/google-signin": "^12.1.0",
+    "@react-navigation/native": "^6.1.17",
+    "@react-navigation/native-stack": "^6.9.26",
     "expo": "~51.0.8",
     "expo-status-bar": "~1.12.1",
     "firebase": "^10.12.2",
     "react": "18.2.0",
     "react-native": "0.74.1",
     "react-native-google-signin": "^2.1.1",
-    "request": "^2.88.2",
+    "react-native-safe-area-context": "4.10.1",
     "react-native-screens": "3.31.1",
-    "react-native-safe-area-context": "4.10.1"
+    "react-navigation": "^5.0.0",
+    "request": "^2.88.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
