+++ screens/Login.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from 'react';
 import { StyleSheet, Text, View, Button } from 'react-native';
-import './firebaseConfig';
+import '../firebaseConfig';
 // maybe you need this...
 //
