+++ screens/components/index.js
@@ -1,0 +1,2 @@
+export { default as Navbar } from './Navbar';
+export { default as Timer } from './Timer';
