+++ screens/index.js
@@ -1,0 +1,3 @@
+export { default as Dashboard } from './Dashboard';
+export { default as Login } from './Login';
+export { default as Welcome } from './Welcome';
